requirements.txt
asgiref==3.7.2
Django==4.2.6
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
phonenumbers==8.13.24
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
